disable backspace fix
No longer needed for most users. Of course if you want to actually use the program in it's current state you can enable this and follow the guide in issue#1 but I don't recommend using the program until I update the way it receives characters (probably just going to use input()). Only recommended use for this is if you have ^H set to backspace in your config.